requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
fastapi==0.115.4
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.41.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.32.0
    # via -r requirements.in
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
